requirements.txt
# Flask and web framework
Flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20

# Object storage
minio==7.1.17

# Deep learning
torch>=1.13.0
torchvision>=0.14.0
timm>=0.9.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
python-dotenv==1.0.0
tqdm>=4.65.0

# API documentation (optional)
flasgger==0.9.7.1
